refactor(@angular/ssr): move `ng-add` schematic to `@schematics/angular`
This move is in preparation to enable `ng new --ssr`.
