+++ packages/angular/ssr/schematics/collection.json
@@ -2,7 +2,7 @@
   "schematics": {
     "ng-add": {
       "description": "Adds Angular SSR to the application without affecting any templates",
-      "factory": "./ng-add",
+      "factory": "./ng-add/index",
       "schema": "./ng-add/schema.json"
     }
   }
