+++ packages/schematics/angular/ssr/files/application-builder/server.ts.template
