+++ packages/schematics/angular/ssr/files/server-builder/server.ts.template
