+++ packages/schematics/angular/ssr/schema.json
@@ -1,6 +1,6 @@
 {
   "$schema": "http://json-schema.org/draft-07/schema",
-  "$id": "SchematicsAngularNgAddSSR",
+  "$id": "SchematicsAngularSSR",
   "title": "Angular SSR Options Schema",
   "type": "object",
   "properties": {
