+++ packages/schematics/angular/utility/latest-versions/package.json
@@ -3,6 +3,7 @@
   "comment": "This file is needed so that dependencies are synced by Renovate.",
   "private": true,
   "dependencies": {
+    "@types/express": "^4.17.17",
     "@types/jasmine": "~4.3.0",
     "@types/node": "^16.11.7",
     "express": "^4.18.2",
